backend/requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.3.1

opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2

ultralytics==8.3.0

# Torch (CPU stable version)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121